fix(payments): corregir declaración de tipo float → int en PaymentQueryService
CIERRE DE ERROR DE CI (exit code 255):
La implementación `PaymentQueryService::getWaiterTipsInSession()` todavía
declaraba `: float` como tipo de retorno, pero la interfaz
`PaymentQueryServiceInterface` fue actualizada a `: int` en P2-005.

Esto causaba un error fatal de PHP:
"Declaration must be compatible with PaymentQueryServiceInterface"

SOLUCIÓN:
Corregir todas las declaraciones de retorno `float` a `int` en
`PaymentQueryService.php` para que coincidan con el contrato de la interfaz.

IMPACTO:
✅ CI pasa exitosamente sin errores de declaración.
✅ Coherencia total entre contratos (int) e implementaciones (int).
✅ Cumplimiento estricto de ADR-018 (Integer CLP).

VERIFICACIÓN:
✅ Tests de Payments pasando (94 tests, 239 assertions).
✅ Suite completa de backend sin errores.

Refs: P2-005 Contratos financieros float → int, ADR-018 Money Contract

+++ app/Modules/Payments/Application/Services/PaymentQueryService.php
@@ -43,7 +43,7 @@ class PaymentQueryService implements PaymentQueryServiceInterface
             ->keyBy('method_code');
     }
 
-    public function getWaiterTipsInSession(int $cashSessionId, int $waiterId): float
+    public function getWaiterTipsInSession(int $cashSessionId, int $waiterId): int
     {
         $this->validateSessionOwnership($cashSessionId);
